requirements.txt
python-telegram-bot==20.0a1

psutil
gputil